requirements.txt
python-binance>=1.0.16
requests==2.28.1
pandas==1.5.2
numpy==1.23.5
sqlalchemy==1.4.45
pytest==7.2.0
pyyaml==6.0
cryptofeed==2.3.0
scikit-learn==1.2.0
matplotlib==3.6.2
pyupbit==0.2.33
dash==2.9.3
plotly==5.14.1
dash-bootstrap-components==1.4.1


